+++ src/admin/pages/ContentType/Edit/CreateField/field-types/CreateChoice/types.ts
