+++ src/admin/pages/ContentType/Edit/CreateField/field-types/DateTime/index.tsx
@@ -1,10 +1,10 @@
-import { SettingsEthernetOutlined } from '@mui/icons-material';
+import { DateRangeOutlined } from '@mui/icons-material';
 import { Accordion, AccordionSummary, Box, Stack, Typography } from '@mui/material';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { Props } from './types';
+import { Props } from '../types';
 
-const SelectDropdownM2oType: React.FC<Props> = (props) => {
+export const DateTimeType: React.FC<Props> = (props) => {
   const { expanded, handleChange } = props;
   const { t } = useTranslation();
 
@@ -32,20 +32,16 @@ const SelectDropdownM2oType: React.FC<Props> = (props) => {
         expanded={expanded}
         square
         disableGutters
-        onChange={() => handleChange('selectDropdownM2o')}
+        onChange={() => handleChange('dateTime')}
       >
         <AccordionSummary aria-controls="panel-content" id="panel-header">
           <Stack direction="row" columnGap={2}>
             <Box display="flex" alignItems="center">
-              <SettingsEthernetOutlined />
+              <DateRangeOutlined />
             </Box>
             <Stack direction="column">
-              <Typography variant="subtitle1">
-                {t('field_interface.select_dropdown_m2o')}
-              </Typography>
-              <Typography variant="caption">
-                {t('field_interface.select_dropdown_m2o_caption')}
-              </Typography>
+              <Typography variant="subtitle1">{t('field_interface.date_time')}</Typography>
+              <Typography variant="caption">{t('field_interface.date_time_caption')}</Typography>
             </Stack>
           </Stack>
         </AccordionSummary>
@@ -53,5 +49,3 @@ const SelectDropdownM2oType: React.FC<Props> = (props) => {
     </Stack>
   );
 };
-
-export default SelectDropdownM2oType;
