+++ src/admin/pages/ContentType/Edit/CreateField/field-types/ListO2m/index.tsx
@@ -1,10 +1,10 @@
-import { InsertDriveFileOutlined } from '@mui/icons-material';
+import { SettingsEthernetOutlined } from '@mui/icons-material';
 import { Accordion, AccordionSummary, Box, Stack, Typography } from '@mui/material';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { Props } from './types';
+import { Props } from '../types';
 
-const FileType: React.FC<Props> = (props) => {
+export const ListO2mType: React.FC<Props> = (props) => {
   const { expanded, handleChange } = props;
   const { t } = useTranslation();
 
@@ -28,15 +28,15 @@ const FileType: React.FC<Props> = (props) => {
           <Typography variant="h6">coming soon..</Typography>
         </Box>
       </Box>
-      <Accordion expanded={expanded} square disableGutters onChange={() => handleChange('file')}>
+      <Accordion expanded={expanded} square disableGutters onChange={() => handleChange('listO2m')}>
         <AccordionSummary aria-controls="panel-content" id="panel-header">
           <Stack direction="row" columnGap={2}>
             <Box display="flex" alignItems="center">
-              <InsertDriveFileOutlined />
+              <SettingsEthernetOutlined />
             </Box>
             <Stack direction="column">
-              <Typography variant="subtitle1">{t('field_interface.file')}</Typography>
-              <Typography variant="caption">{t('field_interface.file_caption')}</Typography>
+              <Typography variant="subtitle1">{t('field_interface.list_o2m')}</Typography>
+              <Typography variant="caption">{t('field_interface.list_o2m_caption')}</Typography>
             </Stack>
           </Stack>
         </AccordionSummary>
@@ -44,5 +44,3 @@ const FileType: React.FC<Props> = (props) => {
     </Stack>
   );
 };
-
-export default FileType;
