+++ src/admin/pages/ContentType/Edit/CreateField/field-types/ListO2o/index.tsx
@@ -1,10 +1,10 @@
-import { PhotoOutlined } from '@mui/icons-material';
+import { SettingsEthernetOutlined } from '@mui/icons-material';
 import { Accordion, AccordionSummary, Box, Stack, Typography } from '@mui/material';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { Props } from './types';
+import { Props } from '../types';
 
-const FileImageType: React.FC<Props> = (props) => {
+export const ListO2oType: React.FC<Props> = (props) => {
   const { expanded, handleChange } = props;
   const { t } = useTranslation();
 
@@ -28,20 +28,15 @@ const FileImageType: React.FC<Props> = (props) => {
           <Typography variant="h6">coming soon..</Typography>
         </Box>
       </Box>
-      <Accordion
-        expanded={expanded}
-        square
-        disableGutters
-        onChange={() => handleChange('fileImage')}
-      >
+      <Accordion expanded={expanded} square disableGutters onChange={() => handleChange('listO2o')}>
         <AccordionSummary aria-controls="panel-content" id="panel-header">
           <Stack direction="row" columnGap={2}>
             <Box display="flex" alignItems="center">
-              <PhotoOutlined />
+              <SettingsEthernetOutlined />
             </Box>
             <Stack direction="column">
-              <Typography variant="subtitle1">{t('field_interface.file_image')}</Typography>
-              <Typography variant="caption">{t('field_interface.file_image_caption')}</Typography>
+              <Typography variant="subtitle1">{t('field_interface.list_o2o')}</Typography>
+              <Typography variant="caption">{t('field_interface.list_o2o_caption')}</Typography>
             </Stack>
           </Stack>
         </AccordionSummary>
@@ -49,5 +44,3 @@ const FileImageType: React.FC<Props> = (props) => {
     </Stack>
   );
 };
-
-export default FileImageType;
