+++ src/admin/pages/ContentType/Edit/EditField/field-types/InputMultiline/index.tsx
@@ -12,15 +12,15 @@ import Grid from '@mui/material/Unstable_Grid2/Grid2';
 import React, { useEffect } from 'react';
 import { Controller, SubmitHandler, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
-import logger from '../../../../../../utilities/logger';
-import { shallowEqualObject } from '../../../../../../utilities/shallowEqualObject';
+import logger from '../../../../../../../utilities/logger';
+import { shallowEqualObject } from '../../../../../../../utilities/shallowEqualObject';
 import updateFieldSchema, {
   FormValues,
-} from '../../../../../fields/schemas/collectionFields/updateField';
-import { useField } from '../Context';
-import { Props } from './types';
+} from '../../../../../../fields/schemas/collectionFields/updateField';
+import { useField } from '../../Context';
+import { Props } from '../types';
 
-const InputMultilineType: React.FC<Props> = (props) => {
+export const InputMultilineType: React.FC<Props> = (props) => {
   const { field: meta, onEditing, onSuccess } = props;
   const { t } = useTranslation();
   const { updateField } = useField();
@@ -146,5 +146,3 @@ const InputMultilineType: React.FC<Props> = (props) => {
     </Stack>
   );
 };
-
-export default InputMultilineType;
